+++ src/main/java/ro/gal/perfectnumber/service/PerfectNumberService.java
@@ -1,8 +1,9 @@
 package ro.gal.perfectnumber.service;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface PerfectNumberService {
-    boolean isPerfectNumber(long number);
+    boolean isPerfectNumber(BigInteger number);
     List<Long> generatePerfectNumbers(long start, long end);
 }
